+++ src/Core/Dom/DomNodeInterface.php
@@ -10,6 +10,7 @@ interface DomNodeInterface
 
     public function hasClass($className);
     public function hasClasses(array $classNames);
+    public function hasAnyClass(array $classNames);
     public function getAttribute($name);
     public function getTagName();
     public function getNodeValue();